vision/requirements.txt
google-cloud-storage >= 1.14.0
google-cloud-automl >= 0.9.0